# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Work Agent",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/proposals": {
      "get": {
        "tags": [
          "proposals"
        ],
        "summary": "Proposals",
        "operationId": "proposals_api_v1_proposals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalView"
                  },
                  "type": "array",
                  "title": "Response Proposals Api V1 Proposals Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Create Proposal",
        "operationId": "create_proposal_api_v1_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/{proposal_id}": {
      "put": {
        "tags": [
          "proposals"
        ],
        "summary": "Edit Proposal",
        "operationId": "edit_proposal_api_v1_proposals__proposal_id__put",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/{proposal_id}/approve": {
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Approve Proposal",
        "operationId": "approve_proposal_api_v1_proposals__proposal_id__approve_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/{proposal_id}/reject": {
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Reject Proposal",
        "operationId": "reject_proposal_api_v1_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Aufgaben und Kalender"
        ],
        "summary": "Tasks",
        "operationId": "tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskView"
                  },
                  "title": "Response Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calendar": {
      "get": {
        "tags": [
          "Aufgaben und Kalender"
        ],
        "summary": "Calendar",
        "operationId": "calendar_api_v1_calendar_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarView"
                  },
                  "title": "Response Calendar Api V1 Calendar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionPayload": {
        "properties": {
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "recipient": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient"
          },
          "due": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "attendees": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attendees"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ActionPayload"
      },
      "ActionResult": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ActionResult"
      },
...
      "CalendarView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "thread_key": {
            "type": "string",
            "title": "Thread Key"
          },
          "web_url": {
            "type": "string",
            "title": "Web Url"
          },
          "meta": {
            "$ref": "#/components/schemas/ItemMetadata"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "processing": {
            "type": "string",
            "title": "Processing"
          },
          "processing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "external_id",
          "kind",
          "title",
          "body",
          "sender",
          "occurred_at",
          "thread_key",
          "web_url",
          "meta",
          "content_hash",
          "available",
          "status",
          "summary",
          "version",
          "processing",
          "source_name",
          "source_kind",
          "ai_enabled"
        ],
        "title": "CalendarView",
        "description": "Calendar metadata includes start/end with source timezone."
      },
...
      "Citation": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "locator": {
            "type": "string",
            "title": "Locator",
            "default": "Original"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          }
        },
        "type": "object",
        "required": [
          "item_id"
        ],
        "title": "Citation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemMetadata": {
        "properties": {
          "demo": {
            "type": "boolean",
            "title": "Demo",
            "default": false
          },
          "start": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceTime"
              },
              {
                "type": "null"
              }
            ]
          },
          "end": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceTime"
              },
              {
                "type": "null"
              }
            ]
          },
          "due": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceTime"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ItemMetadata"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "Ok"
      },
      "ProposalInput": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "payload"
        ],
        "title": "ProposalInput"
      },
      "ProposalView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "payload": {
            "$ref": "#/components/schemas/ActionPayload"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "approved_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Hash"
          },
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "payload",
          "citations",
          "version",
          "status",
          "result",
          "created_at"
        ],
        "title": "ProposalView"
      },
...
      "SourceTime": {
        "properties": {
          "dateTime": {
            "type": "string",
            "title": "Datetime"
          },
          "timeZone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "dateTime"
        ],
        "title": "SourceTime"
      },
...
      "TaskView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "thread_key": {
            "type": "string",
            "title": "Thread Key"
          },
          "web_url": {
            "type": "string",
            "title": "Web Url"
          },
          "meta": {
            "$ref": "#/components/schemas/ItemMetadata"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "processing": {
            "type": "string",
            "title": "Processing"
          },
          "processing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "external_id",
          "kind",
          "title",
          "body",
          "sender",
          "occurred_at",
          "thread_key",
          "web_url",
          "meta",
          "content_hash",
          "available",
          "status",
          "summary",
          "version",
          "processing",
          "source_name",
          "source_kind",
          "ai_enabled"
        ],
        "title": "TaskView",
        "description": "Task metadata includes due date and source status."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionInput": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "VersionInput"
      },
...
    }
  }
}